Splitting migrations
